refactor: Modularize
