+++ src/types/element.ts
@@ -1,0 +1,5 @@
+export interface Element {
+  name: string;
+  xpath: string;
+  url: string;
+}
