+++ src/types/result.ts
@@ -1,0 +1,9 @@
+interface ScrapeResult {
+  xpath: string;
+  text: string;
+  name: string;
+}
+
+export type Result = {
+  [key: string]: ScrapeResult[];
+};
